Remove occurrences of exported Terraform version
